==> 1.claculate-closest-distance-from-fountain-furniture.ipynb <==
import os
import folium
import numpy as np
import pandas as pd
import geopandas as gpd
import matplotlib.pyplot as plt
from shapely.geometry import Polygon

from utils import *

### Fountains

fountains = pd.read_csv(os.getenv("fountains"))
fountains.columns = fountains.columns.str.lower()

fountains = gpd.GeoDataFrame(fountains, geometry=gpd.points_from_xy(fountains.longitude, fountains.latitude))

print(fountains.crs)

fountains = fountains.set_crs('epsg:4326')

print(fountains.crs)

columns = ["precision_localisation", "date_installation", "etat", "x", "y"]
fountains = fountains.drop(columns=columns)

fountains.head(1)

### Urban furniture

mobilier = gpd.read_file(os.getenv("mobilier"))
mobilier.columns = mobilier.columns.str.lower()

print(mobilier.crs)

mobilier = mobilier.to_crs("epsg:4326")

print(mobilier.crs)

columns = ["element", "inspectdat", "actif", "ancree", "instaldate", "materiau", "grarg",
           "grasec", "indexgparc", "typeancrag", "modele", "x", "y"]
mobilier = mobilier.drop(columns=columns)

mobilier.head(1)

### calcul the colsest distance between fountain and urban furniture 

fountains_utm = fountains.to_crs(epsg=32188)
mobilier_utm = mobilier.to_crs(epsg=32188)

print(fountains_utm.crs)
print(mobilier_utm.crs)

fountain_w_furniture_data = fountains_utm.sjoin_nearest(mobilier_utm, distance_col="nearest_urban")

min(fountain_w_furniture_data["nearest_urban"])

max(fountain_w_furniture_data["nearest_urban"])

m = folium.Map(location=[45.5414195, -73.6303031], tiles="cartodbpositron", zoom_start=12)

part = fountain_w_furniture_data[fountain_w_furniture_data.nom_parc_lieu.str.contains("Royal", na=False)]
for _, row in part.iterrows():
    lat = row.latitude
    lon = row.longitude
    folium.Marker(location=[lat, lon], popup=row["id"]).add_to(m)

m

part = fountain_w_furniture_data[fountain_w_furniture_data.nom_parc_lieu.str.contains("Royal", na=False)]

for _, row in part.iterrows():
    lat = row.lat
    lon = row.long
    folium.CircleMarker(
        location=[lat, lon], 
        radius=2,
        color="red",
        fill=True,
        fill_opacity=0.7,
        fill_color="red",
        popup=row["objectid"]).add_to(m)

m
