==> 8.counts-fountains-in-parcs.ipynb <==
import os
import folium
import requests
import numpy as np
import pandas as pd
import geopandas as gpd
import matplotlib.pyplot as plt
from shapely.geometry import Polygon

from utils import *

parcs_url="https://donnees.montreal.ca/dataset/2e9e4d2f-173a-4c3d-a5e3-565d79baa27d/resource/35796624-15df-4503-a569-797665f8768e/download/espace_vert.json"
fountaines_url="https://donnees.montreal.ca/dataset/3ff400f3-63cd-446d-8405-842383377fb8/resource/26659739-540d-4fe2-8107-5f35ab7e807c/download/fontaine_eau_potable_v2018.csv"


WGS84=4326
MTM8=32188

def process_request(url):
    response = requests.get(url)
    response.raise_for_status()
    return response.json()


def json_items_to_geodataframe(http_response):
    return gpd.GeoDataFrame().from_features(http_response).set_crs(f"epsg:{WGS84}")



def process_fountains_data(url):
    df = pd.read_csv(url)
    return df


def create_fountains_geodataframe(df):
    gdf = gpd.GeoDataFrame(df, geometry=gpd.points_from_xy(df.Longitude, df.Latitude)).set_crs(f"epsg:{WGS84}")
    return gdf

fountaines_df = process_fountains_data(fountaines_url)
fountaines_gdf = create_fountains_geodataframe(fountaines_df)

http_response = process_request(parcs_url)
parcs_gdf = json_items_to_geodataframe(http_response)

print(fountaines_gdf.crs)
print(parcs_gdf.crs)

parcs_gdf.columns = parcs_gdf.columns.str.lower()
fountaines_gdf.columns = fountaines_gdf.columns.str.lower()

columns = ["id", "longitude", "latitude", "geometry"]
fountaines_gdf = fountaines_gdf.filter(columns)

columns = ["objectid", "superficie", "geometry"]
parcs_gdf = parcs_gdf.filter(columns)

fountaines_in_parcs = fountaines_gdf.sjoin(parcs_gdf, predicate='within').drop(columns="index_right")
fountaines_in_parcs = fountaines_in_parcs.astype({
                        "id": "int64", 
                        "objectid": "int64",
                        "longitude": "float64",
                        "latitude": "float64",
                        "superficie": "float64"})

# group by objectid and count rows
grouped = fountaines_in_parcs.groupby('objectid').count()
grouped

# add new column 'count' to original dataframe
fountaines_in_parcs['count'] = fountaines_in_parcs['objectid'].map(grouped['id'])

fountaines_in_parcs

fountaines_in_parcs.drop(columns="geometry").to_csv("fountains-in-parcs-counts.csv", index=False)

import pandas as pd

# create example dataframe
df = pd.DataFrame({
    'id': [1, 1, 2, 2, 2, 3],
    'col1': ['a', 'b', 'c', 'd', 'e', 'f'],
    'col2': [10, 20, 30, 40, 50, 60]
})

# group by ID and count rows
grouped = df.groupby('id').count()

# add new column 'count' to original dataframe
df['count'] = df['id'].map(grouped['col1'])

# display the result
print(df)


grouped
