==> 9.join-distances-and-fountains-counts.ipynb <==
import os
import folium
import numpy as np
import pandas as pd
import geopandas as gpd

fountains_distances = pd.read_csv("fountains-nearest-distances.csv")
fountains_counts = pd.read_csv("fountains-in-parcs-counts.csv")

print(f"fountains distances: {len(fountains_distances)}")
print(f"fountains counts: {len(fountains_counts)}")

fountains_distances.head(2)

fountains_counts.head(2)

fountains_distances = fountains_distances.astype({
                            "id": "int64", 
                            "nearest_rev": "float64",
                            "nearest_fountain": "float64",
                            "nearest_rvertes": "float64",
                            "nearest_pcyclable": "float64",
                            "nearest_murbain": "float64", 
                            "nearest_iexterne": "float64",
                            "min_col": "str",
                            "min_val": "float64"})

fountains_counts = fountains_counts.astype({
                            "id": "int64", 
                            "longitude": "float64",
                            "latitude": "float64",
                            "objectid": "int64",
                            "superficie": "float64",
                            "count": "int64"})

inner_join = fountains_distances.merge(fountains_counts, on="id", how="inner")

print(f"inner join: {len(inner_join)}")

inner_join

inner_join.to_csv("fountains-distances-join-fountains-counts.csv", index=False)
